+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
-### Next
+### 0.3.0
 - Added support for Pedometer sensor.
 - Added support for Altimeter sensor.
-- Added isSensorModuleAvailable method to check for the availability of the sensor in the device.
+- Added isSensorModuleAvailable: method to check for the availability of the sensor in the device.
+- Added csvHeaderForSensorModule: method to get the headers of the csv format.
+- Improved csvString by providing the timestamp in both string and timeIntervalSince1970 format.
+- Added SKSensorTimestamp class for better managing sensor timestamps.
+- Improved dictionaryData, making it compatible with JSON format.
 
 ### 0.2.0 (August 23, 2015)
 - Added support for Eddystone™ Proximity sensor.